% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA22231079.R
\name{RK4}
\alias{RK4}
\title{Solve for first order ODEs by R-K 4th order}
\usage{
RK4(dynamics, y0, t0 = 0, tn = 1, n = 1000)
}
\arguments{
\item{dynamics}{the function of t and y in the explicit ODE \eqn{y'=f(t,y)}}

\item{y0}{the initial value}

\item{t0}{the initial value input}

\item{tn}{the terminal value input}

\item{n}{the number of sub-intervals to use, defaults to 1000}
}
\value{
data.frame of the time grid and solution grid
}
\description{
{One dimensional first order ODE solver using either RK4}
}
\examples{
\dontrun{
  f1 <- function(t, v) 9.8 - v^2
  f2 <- function(t, v) 9.8 - 2*v^2
  f3 <- function(t, v) 9.8 - 3*v^2
  v1 <- RK4(dynamics = f1, y0 = 0, tn = 1, n = 5000)
  v2 <- RK4(dynamics = f2, y0 = 0, tn = 1, n = 5000)
  v3 <- RK4(dynamics = f3, y0 = 0, tn = 1, n = 5000)
  plot(v1, col = 'red')
  par(new=TRUE)
  plot(v2, col = 'blue')
  par(new=TRUE)
  plot(v3, col = 'green')
}
}
